Fix full disc scanning
In Media object initialization, if file path isn't a directory and title is nil, set title to 0. This is the case of --scan on an iso rip.

Don't try to evaluate media.info for debugging in --scan logic if title is nil has this causes an early exit in .info since there are multiple titles.